Move accessor methods to production code
