+++ pkg/sdk/client.go
@@ -52,6 +52,14 @@ type Client struct {
 	Warehouses       Warehouses
 }
 
+func (c *Client) GetAccountLocator() string {
+	return c.accountLocator
+}
+
+func (c *Client) GetConfig() *gosnowflake.Config {
+	return c.config
+}
+
 func NewDefaultClient() (*Client, error) {
 	return NewClient(nil)
 }
